+++ src/Stack/Options/HpcReportParser.hs
@@ -14,6 +14,7 @@ hpcReportOptsParser = HpcReportOpts
     <$> many (textArgument $ metavar "TARGET_OR_TIX")
     <*> switch (long "all" <> help "Use results from all packages and components involved in previous --coverage run")
     <*> optional (strOption (long "destdir" <> help "Output directory for HTML report"))
+    <*> switch (long "open" <> help "Open the report in the browser")
 
 pvpBoundsOption :: Parser PvpBounds
 pvpBoundsOption =